requirements.txt
Python #3.8 or higher(3.9 preferred)
PyTorch
OpenCV
Ultralytics 
#Install the requirements using one command 
#pip install -r requirements.txt
